api/requirements.txt
fastapi==0.109.2
uvicorn==0.27.1
python-dotenv==1.0.1
psycopg2-binary==2.9.9
sqlalchemy==2.0.27
python-jose[cryptography]==3.3.0
passlib==1.7.4
bcrypt==3.2.2
python-multipart==0.0.9
httpx==0.26.0
stripe==7.11.0
google-auth==2.28.1
pydantic[email]==2.6.1
alembic==1.13.1
slowapi==0.1.9
beautifulsoup4==4.13.3
lxml==5.3.1
pandas==2.2.3
resend==2.7.0